+++ src/services/ForecastService.ts
@@ -1,19 +1,19 @@
 import axios from 'axios';
-import dayjs from 'dayjs';
 
-class WeatherService {
+interface ForecastRequest {
+  baseDate: string;
+  baseTime: string;
+  x: number;
+  y: number;
+}
+
+class ForecastService {
   private BASE_URL: string =
     'http://apis.data.go.kr/1360000/VilageFcstInfoService_2.0';
 
   // 기상청 초단기예보조회 서비스
-  public getUltraShortTermForecast(data: { x: number; y: number }) {
-    const date = new Date();
-
-    // 기상청 API 제공 시간에 따라 기준 시간 변경
-    if (date.getMinutes() < 45) date.setHours(date.getHours() - 1);
-
-    const baseDate = dayjs(date).format('YYYYMMDD');
-    const baseTime = dayjs(date).format('HH30');
+  public getUltraShortTermForecast(data: ForecastRequest) {
+    const { baseDate, baseTime, x, y } = data;
 
     return axios.get(`${this.BASE_URL}/getUltraSrtFcst`, {
       params: {
@@ -23,19 +23,15 @@ class WeatherService {
         dataType: 'JSON',
         base_date: baseDate,
         base_time: baseTime,
-        nx: data.x,
-        ny: data.y,
+        nx: x,
+        ny: y,
       },
     });
   }
 
   // 기상청 단기예보조회 서비스
-  public getShortTermForecast(data: { x: number; y: number }) {
-    const date = new Date();
-    date.setDate(date.getDate() - 1);
-
-    const baseDate = dayjs(date).format('YYYYMMDD');
-    const baseTime = dayjs(date).format('2300');
+  public getShortTermForecast(data: ForecastRequest) {
+    const { baseDate, baseTime, x, y } = data;
 
     return axios.get(`${this.BASE_URL}/getVilageFcst`, {
       params: {
@@ -45,11 +41,11 @@ class WeatherService {
         dataType: 'JSON',
         base_date: baseDate,
         base_time: baseTime,
-        nx: data.x,
-        ny: data.y,
+        nx: x,
+        ny: y,
       },
     });
   }
 }
 
-export default new WeatherService();
+export default new ForecastService();
